Component INspector base workign
Now I can properly add the inspector for every component I keep on making.

I don't know if direct acces through externs is a good way but it is very handy.

It is not, but having copy on data is so not memory efficient and I have to send the data and it would be slowwww... for now I will not bother and keep on getting things.