Adding messaging system for better control
A simple messaging system that should be polymorpher and adapted to  declare some events that each messager can send.

Then a 2nd step to subscribe to any kind of message type (which could still be called before exection) It is flexible in the sense that could be define on scripting code but could nto be used by engine could then.

Engine code can be subscribed by scripting code.