Changed to Templating
Some things are way nicer and have more control as well as being easier to read and write. The issue now is that is Inspector sitll has to use  Switch to add the types while having variant memory.

With teh class I now get control to when the data changes are applied and can be controlled with an inspector function to set when that happens (after all updates, after each change, as events,... everything should be doable!).