Small rewrites for linux
Linux differenciates strictly between R References and L References, which ones allow passage of local memory and the other doesn't. THat was throwing an error on some returns that were already bad. Reference returns for invalid values now returnto a static invalid data value, that can be made sure that is an invalid one.